Optimize PPDF pipeline: sparse path lengths, float32, vectorized ops
Resolves 12 performance and correctness issues in the 3D raytracing/PPDF pipeline:

Severe:
- Replace dense (N_rays × N_obj) L_objects allocation with SparsePathLengths
  dataclass; narrow-phase intersections computed only for broad-phase hits,
  accumulation via index_add_ — O(N_hits) memory instead of O(N_rays × N_obj)
- Vectorize triplet extraction in system matrix assembly: nonzero() + tensor
  gather replaces nested Python loop with per-element .item() calls

High:
- Avoid double ray computation per chunk: pre-computed o/d tensors passed from
  broad phase into ppdf_3d_local via new o_world/d_world optional params

Medium:
- Remove redundant internal chunking loop from build_ray_object_candidate_lists;
  single vectorized ray_aabb_intersect call replaces while-loop
- Remove unnecessary O(N log N) argsort on broad-phase pairs (nonzero output
  is already row-major sorted)
- Switch global DTYPE from float64 to float32 throughout pipeline (config,
  broad phase, narrow phase, solid angle, PPDF accumulation) — halves memory,
  improves GPU throughput
- Replace Python loop for dense-row assembly with torch.scatter_ in both
  entry scripts

Low (correctness):
- Replace fragile half_sizes[global_obj_ids] OBB indexing with explicit
  global_to_obb_local reverse-lookup map — correct regardless of object ordering
- Replace brittle poly_local = global_id - N_obb arithmetic with explicit
  global_to_poly_local reverse-lookup map
- Expand solid angle to all 5 detector faces (front + 4 sides, excluding back
  face identified by max radial distance) to capture oblique-incidence photon
  entry; both entry scripts updated to build 5-face voxel_faces tensors

+++ scanner_modeling/_config.py
@@ -1,7 +1,7 @@
-from torch import float64
+from torch import float32
 
 #: Global floating-point dtype for all 3D geometry / raytracing tensors.
-DTYPE = float64  # fixed precision
+DTYPE = float32
 
 #: Canonical distance unit used throughout geometry/raytracer.
 #: All distances passed into the 3D pipeline must be in these units.
